item-or-inner-list into inner-list
