+++ http_sfv/dictionary.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
 from http_sfv.item import ser_item, bin_ser_item
-from http_sfv.list import parse_item_or_inner_list, bin_parse_item_or_inner_list
+from http_sfv.innerlist import parse_item_or_inner_list, bin_parse_item_or_inner_list
 from http_sfv.parameters import parse_params, ser_params
 from http_sfv.types import DictionaryType
 from http_sfv.util import (
