+++ http_sfv/list.py
@@ -1,26 +1,18 @@
-from typing import Tuple, Union, List, cast
+from typing import Tuple
 
 from http_sfv.innerlist import (
-    parse_innerlist,
-    bin_parse_innerlist,
-    ser_innerlist,
-    bin_ser_innerlist,
+    parse_item_or_inner_list,
+    bin_parse_item_or_inner_list,
+    ser_item_or_inner_list,
+    bin_ser_item_or_inner_list,
 )
-from http_sfv.item import (
-    PAREN_OPEN,
-    parse_item,
-    bin_parse_item,
-    ser_item,
-    bin_ser_item,
-)
-from http_sfv.types import ListType, ItemType, InnerListType, ItemOrInnerListType
+from http_sfv.types import ListType
 from http_sfv.util import discard_http_ows
 from http_sfv.util_binary import (
     encode_integer,
     decode_integer,
     bin_header,
     STYPE,
-    HEADER_OFFSET,
 )
 
 
@@ -67,36 +59,3 @@ def bin_ser_list(_list: ListType) -> bytes:
     for member in _list:
         data.append(bin_ser_item_or_inner_list(member))
     return b"".join(data)
-
-
-def parse_item_or_inner_list(data: bytes) -> Tuple[int, Union[ItemType, InnerListType]]:
-    try:
-        if data[0] == PAREN_OPEN:
-            return parse_innerlist(data)
-    except IndexError:
-        pass
-    return parse_item(data)
-
-
-def bin_parse_item_or_inner_list(
-    data: bytes, cursor: int
-) -> Tuple[int, Union[ItemType, InnerListType]]:
-    if data[0] >> HEADER_OFFSET == STYPE.INNER_LIST:
-        return bin_parse_innerlist(data, cursor)
-    return bin_parse_item(data, cursor)
-
-
-def ser_item_or_inner_list(thing: ItemOrInnerListType) -> str:
-    if not isinstance(thing, tuple):
-        thing = cast(ItemType, (thing, {}))
-    if isinstance(cast(InnerListType, thing)[0], List):
-        return ser_innerlist(cast(InnerListType, thing))
-    return ser_item(cast(ItemType, thing))
-
-
-def bin_ser_item_or_inner_list(thing: ItemOrInnerListType) -> bytes:
-    if not isinstance(thing, tuple):
-        thing = cast(ItemType, (thing, {}))
-    if isinstance(cast(InnerListType, thing)[0], List):
-        return bin_ser_innerlist(cast(InnerListType, thing))
-    return bin_ser_item(cast(ItemType, thing))
